Updated BitCOde Module to prepare support for SharedModule use in OrcJIT
# Breaking Changes
The constructors for BitcodeModule all rewuire a Context. The idea of auto creating the context was convenient at first but is actually an inversion of the real ownership, which made the code more complex for clean up and tragically broken in the face of the ref counted shared module refs for OrcJIT. Thus it is time to unwind that choice.