Xplat build of native code working (#3)
This is a big change because it includes cached binaries and headers from the release versions of Mac and Linux LLVM along with the cached build output for Windows. The default behavior of the pipeline is now to use the cached binaries, resulting in about a 20 min build time, while the full 3 hour build of LLVM is still an option. Lots of clean-up on the scripts is still left to do, as well as creating a multistage pipeline that can actually produce and test the fully formed nuget package, but this is a big chunk of the way there.