Resolves #200
# BREAKING CHANGE
This is a breaking change. The behavior of Context.CreateConstantString(string) has changed slightly. It now constructs a valid C string WITH a null terminator, which is generally what would be expected of something called "string". (The (string,bool) overload remains, to allow apps to be explicit with intent and get the previous behavior) Additionally, the ConstantDataSequential.IsString property now reflects whether the string is a C string (terminating null but no embedded nulls) and the ConstantDataSequential.IsI8Sequence was added to provide the previous behavior of IsString, which was simply that the underlying sequence element type was i8 (with or without a terminator)
---
* Added error handling to parsing of YAML configuration in bindings generator app.
* Added RawData property to ConstandDataSequential (and used it for the ExtractAsString implementations
* Added ExtractAsString overload that accepts an Encoding
* ExtractAsString now automatically discards the terminating null if the underlying data is a C String, while allowing extraction of strings without a terminator and in different encodings.