Updates for simpler native contexts Native contexts are effectively an opaque `void*` they can have any value that fits in a pointer. The native layers don't care about it and store the value to provide as a parameter to callback functions. The implementation of the callback has all the knowledge to interpret the intended meaning of the context value.
For managed code this is an allocated GCHandle converted to a pointer. The managed code can resolve the pointer to a handle, and then the target instance to get a viable managed ref. When no more call backs can occur the context is released. This simplifies the use of callbacks as a native context can be created for any managed heap object. (Extension method added for this action)

* Updated comments
* Removed some dead code
* Added comments to explain limits of the C# 14 keyword `extension` and why it isn't used.
* Removed SafeGCHandle and MarshalGCHandles as context handles are MUCH simpler now.
* Added the memory allocator base types
    - These are not yet used or tested and are subject to change in the future and therfore attributed as experimental.
* Defined interface for materializer callbacks to allow internal implementation to hold delegates as a single unit to become a native context for call backs.
* Removed unnecessary using statements